+++ src/main/java/com/amagari/translationtool/mixin/client/MinecraftMixin.java
@@ -2,6 +2,7 @@ package com.amagari.translationtool.mixin.client;
 
 import com.amagari.translationtool.client.WorldLanguageClient;
 import com.amagari.translationtool.client.WorldLanguageContext;
+import com.amagari.translationtool.client.paratranz.ParaTranzContext;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.server.WorldStem;
@@ -34,5 +35,6 @@ public class MinecraftMixin {
 	@Inject(method = "disconnect(Lnet/minecraft/client/gui/screens/Screen;ZZ)V", at = @At("HEAD"))
 	private void amagari_translation_tool$clearWorldLanguageDirectory(Screen screen, boolean transferring, boolean clearDownloadedPacks, CallbackInfo callbackInfo) {
 		WorldLanguageContext.leaveWorld();
+		ParaTranzContext.resetSessionState();
 	}
 }